requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
autopep8==2.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.19.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
lxml==5.2.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
passlib==1.7.4
playwright==1.43.0
pluggy==1.4.0
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyee==11.1.0
PySocks==1.7.1
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
text-unidecode==1.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wsproto==1.2.0
